Rename AccountReset::GrantRequestsAndSendEmails job
