+++ lib/tasks/account_reset.rake
@@ -3,6 +3,6 @@
 namespace :account_reset do
   desc 'Send Notifications'
   task send_notifications: :environment do
-    AccountReset::GrantRequestsAndSendEmails.new.call
+    GrantAccountResetRequestsAndSendEmails.new.call
   end
 end
